rename PLUT42 auxiliary dir for more readability
